+++ test/system/lua/LuaSystemClass.spec.ts
@@ -175,15 +175,4 @@ import { LuaSystemTestBase } from "./LuaSystemTestBase";
             12, 3, 4
         ]);
     }
-
-    @test "Escape sequences in strings"() {
-
-        const tsCode = `
-            return "A\\b";
-        `;
-
-        expect(this.getLuaParseResult(tsCode)).to.deep.equal([
-            "A\\b"
-        ]);
-    }
 }