+++ test/target/lua/LuaArrayBindingPattern.spec.ts
@@ -43,6 +43,13 @@ import { UnitTest } from "../../UnitTest";
             expected: [
                 "local a, b, _, c = test()"
             ]
+        }, {
+            code: `
+                const [a,b,,c,] = [1,2,3,4,5];
+            `,
+            expected: [
+                "local a, b, _, c = 1, 2, 3, 4, 5"
+            ]
         }])
     }