+++ test/target/wow/WowPathBuilder.spec.ts
@@ -10,25 +10,25 @@ import { UnitTest } from "../../UnitTest";
             code: `import {C} from "./test";`,
             expected: [
                 "local __FILE_META = {...}",
-                "local C = __library.get(\"9606327c7552bfbe0677bea77d8ae226\").C"
+                "local C = __library.get(\"test/index\").C"
             ]
         }, {
             code: `import {D as Q} from "./test/test2";`,
             expected: [
                 "local __FILE_META = {...}",
-                "local Q = __library.get(\"3dea18d42ca0a7008917a4f0c97930fa\").D"
+                "local Q = __library.get(\"test/test2/index\").D"
             ]
         }, {
             code: `import {D} from "@wartoshika/test/test2";`,
             expected: [
                 "local __FILE_META = {...}",
-                "local D = __library.get(\"128b6ada17ddc50f2da51e32a5854e3a\").D"
+                "local D = __library.get(\"node_modules/@wartoshika/test/test2/index\").D"
             ]
         }, {
             code: `import {D} from "@wartoshika/test/test2/../test2";`,
             expected: [
                 "local __FILE_META = {...}",
-                "local D = __library.get(\"128b6ada17ddc50f2da51e32a5854e3a\").D"
+                "local D = __library.get(\"node_modules/@wartoshika/test/test2/index\").D"
             ]
         }], {
                 personalizedLibrary: false